Move: team.services 이름을 더 직관적으로 변경 #36
