+++ src/lib/queries/useTeamQuery.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
-import { fetchGetTeams } from '../services/team.services';
+import { fetchGetTeams } from '../services/team-client.services';
 import { TeamData } from '@/types/teams.type';
 
 // 팀 데이터 가져오기 훅
